requirements.txt

fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
openai>=1.37.0
lxml>=5.2.2
